Revert "[dv] Add a quick exit in csr_rd_sub when in reset"
This reverts commit b400b907d0891418ad4e12e05e6213e68a01391c.

The change there sort of made sense: if there's a reset when we're
doing a CSR operation, it's probably more defensible to return with a
silly answer than to block indefinitely. Unfortunately, this doesn't
quite work with the pwrmgr DV code, which seems to set the under_reset
flag well before it causes a reset (and the code doesn't seem to clear
it again either).

Thinking harder about the problem, I think the correct behaviour
should probably be that the driver/monitor are in charge of responding
quickly when they see a reset on the interface in question. Revert
the (badly thought out) initial change from me in the meantime.

This is a cherry pick of commit deea8d63a86fd085244576b6d363aacd52bac343
to branch earlgrey_1.0.0.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>