[jtag,dv] Fix jtag driver support for trst_n
The existing code was sort of reasonable, but didn't properly handle
the situation where trst_n went low half-way through handling a
request.

It turns out that I was making it much more complicated than it needed
to be. Stop doing that!

This is a cherry pick of commit 423ae18e1d94379bd6d7265cfc980298874c8298
to branch earlgrey_1.0.0.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>