[rv_dm,dv] Weaken check_dmistat when we are in reset
This is needed in order that an injected reset doesn't cause the check
in the sequence to fail. The code also gets a bit longer because it
has a slightly more descriptive error message for a mismatch than the
one you get from DV_CHECK_EQ (including using %02b to print the 2-bit
numbers).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>