[hw,dma,rtl] Flexible address increment options
This PR adds support for a flexible address increment mode.
For the source and destination address, the DMA can be
configured to perform:
* No address increment (FIXED). The DMA always reads/writes
from the same address.
* Wrapped increment (WRAP). The DMA increments and when
finishing one chunk wraps back to the starting address.
* Linear increment (LINEAR). The DMA always increments the
address

These modes are independent of the hardware handshake mode.
It was implemented as two dedicated registers to provide
support for future changes w/o changing the register
layout.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>