[rom_ctrl,dv] Rewrite BusLocalEscChk_A to avoid coverage hole
The "always" in the existing version of this assertion never ran to
completion, so coverage didn't show the assertion as having done
anything.

The assertion is supposed to say "If we get to an Invalid state, we'll
never pass a TL transaction through". Encode this differently by
saying that the Invalid state is terminal and that if we are in that
state then we can't pass a TL transaction through. This way, the
assertion properties will run to completion.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>