[rom_ctrl,dv] Remove irrelevant coverpoint from rom_ctrl_tlul_cg
The check_invalid bin in the coverpoint is not hit, but it doesn't
really matter. The pwrmgr_data_invalid signal means that the
pwrmgr_data_o.good mubi signal that we are generating has an invalid
encoding. This can never happen without an injected fault.

We also don't care about whether this signal is valid or not: it is
not actually read by any code in the IP block, so it doesn’t really
matter if an injected fault causes it to become invalid.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>