[csr_spinwait, dv] csr_spinwait stoppable task
Adding a csr_spinwait which can be stopped by using UVM events.
The old csr_spinwait times out blindly.
This updated csr_spinwait allows for the timeout to stop whenever any
condition such as a SW reset or a module being disabled.

In order to do so the triggering module should use the UVM event pool in
conjunction with the `csr_spinwait_ctrl_object`

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>