[spi_host, configs] Moving clkdivider tests from normal config
clkdiv tests take ages to complete. Potentially over 1h, which is a
waste of time and resources.
In order to quicken up the simulation, the RTL's prescaler decrements
16-units instead of 1 when running the clock divider test.

So there is a base spi_host config, the old config (where the clock
divider tests have been removed from), and a clock divider configs
which forces greater subtractions in the pres-caler.
Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>