[dv] add vseq to check ROM hash in GLS
The final ROM image is built for the "silicon" device, and as such does
not write to the DV_SIM_WINDOW to signal a test pass. This adds a DV
vseq that will signal test pass once the UART prints the expected ROM
hash.

This is a cherry pick of commit 3fa058bde08fa88691fb09e644d805aaef7c972e
to branch earlgrey_1.0.0.

Signed-off-by: Tim Trippel <ttrippel@google.com>