[ci,bazel] Add a script to do a test-tag-filter-like query
The --test_tag_filters option of `bazel test` is very convenient
but is only available for testing and not for querying. This script
takes as input a list of tags in the same fort as test_tag_filters
and runs an equivalent bazel query to list all tests that match.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 31d9369244d9e24630d448da4f900e933bf03657)