[ci] Run airgapped tests with -e and -x flags
Usually bash sub-scripts inherit these flags from their parent shell,
but that doesn't seem to be happening when run via `ip` and `sudo`.

We're seeing cases where the `cquery` fails due to OOM and generates no
tests, but then the `build` succeeds because it only has to build the
hard-coced `mask_rom` target. This success was used as the exit code for
the whole script.

Adding `-e` and `-x` flags ensures the sub-script uses the same flags as
the parent script.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>