[sram_ctrl,dv] Wait with reset after mem_init
Currently, some of the sram_ctrl_stress_all fail with a DataKnown_A assertion.

The reason for this failure is that the sram_ctrl_stress_all test, depending on
the seed, immediately triggers a reset before the first test is started.
However, this reset before the first test interferes with the hardware-based
memory initialization feature.

More specifically, this reset currently happens when the last write of the mem
init is issued, i.e., the write gets not finished. Hence, the memory at the last
address is uninitialized, triggering the DataKnown_A assertions.

To circumvent this issue, this commit adds a wait_clks(2) before the reset.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>