[top_earlgrey] Reduce number of PRINCE half rounds for main SRAM
This commit reduces the number of PRINCE half rounds for the scrambling
of the main SRAM to 2 half rounds as used in the past. This is done
to reduce timing pressure.

This is a cherry pick of commit 0ff694a0812fe2f67da5716385bb3bc7443a9d48
to branch earlgrey_1.0.0.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>