[ottf,rom_e2e] add ability to silience OTTF console prints
The OTTF prints several status messages over the console before and
after a test is executed. This can be noisy for tests that run in sim
and need to monitor the UART TX output. This adds a feature to the OTTF
to silence said messages, and updatest the ROM E2E self hash test to
minimize what is printed over the UART to reduce similation time as this
test is run in GLS where no backdoor logging is enabled.

This is a cherry pick of commit 33c4b71134a6ddbc03bfe3adb852afaa35622c99
to branch earlgrey_1.0.0.

Signed-off-by: Tim Trippel <ttrippel@google.com>