[pentest] Add pentest_call_and_sleep
This commit adds the pentest_call_and_sleep function to the
pentest_lib. This function is similar to the sca_call_and_sleep
function in the sca library but adds the possibility of waiting
until the OTBN execution has finished. Adding this to the
existing function is not possible because this function also needs
to work on english breakfast where OTBN is not available.
!OT_IS_ENGLISH_BREAKFAST did not work as already including
//sw/device/lib/crypto/drivers:otbn into the bazel build rule
raised some compilation errors.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>