[rom] fix ROM_EXT immutable section to work with address translation
The original ROM_EXT immutable section configuration did not account for
the translation addjustments to the immutable section entry point. This
fixes that error and enhances the tests to ensure the same ROM_EXT
mutable section offset can be used across tests that load a ROM_EXT in
different slots (A, B, or Virtual slots).

Signed-off-by: Tim Trippel <ttrippel@google.com>