[mem_bkdr_util] Support varying No. of PRINCE half rounds for scrambling
Previously, all scrambled memory primitives had to use the same number
of PRINCE half rounds which was set to 3. This commit adds a new and
optional argument to the mem_bkdr_util constructor, to allow configuring
different values. If the argument is not provided, the default of 3
PRINCE half rounds is used.

This is a cherry pick of commit 72ea1a12b15cc3eea8b688c24af984611492cb88
to branch earlgrey_1.0.0.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>