[dv,rom_e2e] update expected ROM hash in GLS self hash test
A new ROM was released after #24417 was merged, however the expected
hash was not updated in the vseq used for the GLS test. This updates the
expected hash to match that in the `.c` file counterpart.

This is a cherry pick of commit 466b306e8b4b06a7b93752baf9fa0f605a7f63b2
to branch earlgrey_1.0.0.

Signed-off-by: Tim Trippel <ttrippel@google.com>