[prim,fpv] Increase timeouts in prim_alert_rxtx_async_assert_fpv
These aren't really reasoned carefully, but the existing timeouts gave
a property that was false (and Jasper found a counterexample).

Bumping them up by a cycle stops Jasper from finding a counterexample,
but it doesn't prove the property in 5 minutes either. Hopefully now
true, but we need some follow-up work to speed up the proof (assuming
the property is true) and get proper confidence.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>