[ci] Add a lint to check for unrunnable tests
We have discovered that some unittest have become unrunnable because
we added my mistake dependencies that are marked as only compatible
with the OT platforms. This means that those unittests are then skipped
when running all tests using //...
This commit adds a lint to make sure that all tests are compatible with
the host platform. For efficiency reasons, this lint is run in the sw_test
step. The reason is that running a cquery is quite expensive as it requires
building the graph and even compiling things for certain repositories. By
running it after building the tests, we are sure not to duplicate this work.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>