[hw,uart,rtl] Implement LSIO trigger
In this minimal implementation, the UART core asserts
iff the TX FIFO fill level is smaller than the configured threshold or
the RX FIFO fill level is greater than or equal to the configured
threshold, independent of IRQ enablement. The trigger is flopped
to avoid CDC issues between different clock domains.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>