[spi_host/rtl] Fix `sd_en_o` of `spi_host_fsm`
This partially reverts commit 7eb53da738a770a4cc00153eb7179a2c2d9bdda9,
which introduced a glitch on the `sd_en_o` output of `spi_host_fsm`.
The reversal is partial in order to minimize the netlist diff.  The
`drive_posedge` signal as well as the four `sd_en_ff_q` flops are now
unused.

This is *not* a cherry pick of a commit on `master`, where the glitch is
planned to be fixed with a full reversal of the commit mentioned above.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>