[prim,lint] Remove dead waivers for prim_fifo_*sync.sv
The first can never match (because the subtraction was changed from -3
to -2 in 7cbebc2e690 in June 2021).

The second *has* never matched: I think the variable in question has
been called unused_decval_msb since the initial commit in the
repository.

I think the third matched some width arithmetic that was done on
pointer values until commit 64272d832be replaced the inline logic with
a prim_fifo_sync_cnt instance. (The waiver came from an initial bunch
of waivers in 2020 and the comment doesn't make sense: it looks like
it was accidentally copied from the previous waiver)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>