[dvsim] deterministic report artefact paths
Use git commit hash to group the report artefacts by IP revision. Drop the
report.[html/json] naming and use a timestamp from the time of dvsim job
start. There is no longer any need for moving the artefacts as they are created
with a unique name to avoid clashes.

The previous strategy was to create the artefacts with the label 'latest' then
rename them with a timestamp if and when the next run wanted to write to
the same place. This meant the timestamps were not aligned, rather stamped with
whatever time that individual file/dir was replaced in the subsequent
run. Making it impossible to determine for certain what replaced files
where from which run. If the next run failed part way through then the
"latest" files could be a mix of old and new results. Where blocks are
added or removed there could also be bleed over between runs.

The other issue was that the summary.html files always pointed to
latest. So if renamed and kept they would still not point to the report files
they are summarising.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>