[i2c,dif] Test i2c dif clock stretching cycle bounds
The i2c block requires a minimum value (for the current setup, at least
4 cycles) for the SCL high and low time in order for clock stretching to
function. This functionality was previously added to the DIF, and this
commit updates the relevant unit tests to also check for these
conditional cases. Modified input clock periods are used that have been
calculated to execute minimum boundary branch cases, and a static
assertion before the test ensures that these stay relevant.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>