[otbn,dv] Model a glitch on the STATUS register
This fixes a problem problem that was visible in the otbn_escalate
test. It is caused by the design bug that is tracked in issue #23903.
The one cycle glitch wouldn't be a big deal but we might be really
unlucky and read the register at just the wrong time!

The change in sim.py changes otbnsim to correctly model the design's
behaviour, with a whopping great TODO message that explains how to
undo the change again.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>
(cherry picked from commit 19a6bfc935369dd6a234f59c0f9c093be9cb2fa9)