[entropy_src,verilator] Redcuce entropy consumption on Verilator tests
This commit modifies the `entropy_src_fw_override_test` such that when
running on Verilator, the random number consumer and the AES test utils
are both given 48x larger timeouts, to account for the fact that the
AST's noise source has not been tuned for a Verilator execution
environment. This stops these specific tests failing on Verilator.

Additionally, the requirements of the test have been conditionally
reduced when executing on a Verilator simulation environment. The final
phase of testing is skipped (single bit mode disabled, without an
output delay) as this provides negligible useful information when
running on Verilator, and only the `kDifEntropySingleBitModeDisabled` &
`kDifEntropySingleBitMode0` modes are used to run the tests, as modes
1 through 3 are almost identical and all are much slower than
ModeDisabled. This provides the speedup necessary for the test to
succeed (given reasonable timeout) in the Verilator sim environment.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>