[topgen] Fail more understandably if we start in an unexpected place
This should fix issue #24595. The problem that caused the issue is
that we compute paths to the various IP blocks that should be included
by following relative paths from the `-t` argument. This doesn't work
very well if the top_earlgrey.hjson file moves around!

A sensible design would probably point more explicitly at where to
find IPs, but this should hopefully be a little easier to debug.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>