[pentest] Move SCA lib into pentest lib
Currently, we have the SCA lib that is used for the old simpleserial
based SCA/FI tests (sw/device/sca) and the new pentest lib that is used
for the uJSON based pentest framework (sw/device/tests/penetrationtests).

As in the longerm we want to get rid of the old simpleserial framework,
this commit merges these two libraries and moves it into the new
pentest framework. Moreover, the sca_ suffix is replaced with pentest_.

A cleanup of these functions will be conducted once the simpleserial
framework is removed.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit def97ec)