[top_earlgrey,pinmux] Enable pull-down for TAP strap 0 at reset
TAP strap 0 is sampled after reset (and only once for life cycle states
that are not TEST_UNLOCKED* or RMA).  To ensure it gets sampled as 0
unless driven to 1 from an external source (and specifically that it
gets sampled as 0 when left floating / not connected), this enables the
pull-down of the pad at reset.

This is a cherry pick of commit 78ad89d1aa5f81cca93a0a5c103eec42e37e4759
to branch earlgrey_1.0.0.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>