[topgen] Add target in toplevel BUILD file to autogenerate top DT code
In order to run the dttool using bazel, we need not only to provide
information about the top but also point to the topgen-generated
cfg (top_<name>.gen.hjson) AND the IP block hjson. Unfortunately,
this latter information is currently not recorded anywhere but topgen
knows it so it can render it in the bazel file.

Note that even topgen cannot reliable generate the list of IP DT
dependencies without some hacks because of our very inconsistent
naming scheme. Therefore we for now point to `//hw/<top>:dt_headers`
which is a manually created list.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>