[bazel] Add a new /hw/top directory to point to the actual top
In the future, we want SW to depend on top elements without knowing
which top is it compiled for. The way to achieve this is to introduce
a layer of indirection: SW would depend on e.g. //hw/top:something
instead of //hw/top_earlgrey:something. Currently, this commit
only points everything to earlgrey but more tops can be added in
the future and the top selection becomes part of the bazel
configuration using a config_setting.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>