[sival/aes] Add AES interrupt encryption test
This commit implements the chip_sw_aes_interrupt_encryption test that is
defined in the chip testplan. The purpose of this test is to check,
whether an AES encryption can be interrupted and restored.

Tue Oct 22  Incorporated the changes suggested by reviewers and added AES modes (CFB-128, OFB, CTR) called out in the test plan
Tue Oct 23 Fixed lint errors
Signed-off-by: Ramesh Prakash <rprakas@google.com>