[sw/tests] Add OTP descrambling test
This adds a top-level test for the descrambling functionality/datapath
of `otp_ctrl`.  All checks are done in SW; there's no need for any
checks on the DV side.  The test works by reading out all three SECRET
OTP partitions from SW (which is possible before they get locked) and
confirm that the descrambled value that SW reads from OTP equals the
expected value that is stored in the OTP image (the OTP image itself is
scrambled).

Signed-off-by: Andreas Kurth <adk@lowrisc.org>