[tlul/rtl] Modify readback check error reporting logic
Currently, only a readback error check is reported when the `error_i`
signal is not raised. However, as this signal is also raised when there
is no activity on the TL-UL bus, we might miss some readback failures.

This commit modifies the error reporting logic to also report the
readback mismatch when the error signal is raised.

Closes #24698.

This is a cherry pick of commit 267e61c59160197bd2f659b9f9d17e59b564484a
to branch earlgrey_1.0.0.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>