[aes/dv] Split and clean up aes_message_item string functions
Previously, the basic convert2string() function called in different
contexts (scoreboard, base sequence, etc.) would print information that
is not available / meaningful in the scoreboard context such as the
configuration error data and distributions. This commits splits the
original function into multiple functions to enable controlling what to
print in which context.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>