[ci, sival] Add azure & Github CI testplan validation job on PR
Add an azure CI quick lint job step, and a Github Actions quick
lint job step, which will ensure that testplans are formatted correctly
by running the corresponding validation script, currently set to run on
both the `hw/top_earlgrey/data` and `hw/top_earlgrey/data/ip`
directories. This is also added to the `ci/jobs/quick-lint.sh` job so
that it can easily be run locally.

This is added to both azure and Github Actions as at present both of
these jobs still exist as a result of the partial CI migration from
Azure pipelines to Github Actions.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>