[rom_ctrl,dv] Fixed rom_ctrl_kmac_err_chk_fix failure
The rom_ctrl_kmac_err_chk test fails occasionally as the virtual sequence
ends before the dut has run through it's post-reset flow. This cause a
check in the scoreboard to fail (which is making sure that a test hasn't
finished before it checks anything. The scoreboard executes the check a
cycle earlier when the rom_ctrl_fsm could be in its done state.

The fix adds phase_ready_to_end to the scoreboard and raised an objection
which ensures that run phase doesn't complete until the check has happened.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>