[opentitantool] Improvement to HyperDebug firmware upgrade
Firmware upgrading of HyperDebug requires entering and leaving "DFU
mode", each of which involves resetting connection with the USB device.

Previous code waited one second for the device to reset after upgrading
concluded, before attempting to restablish USB connection with the newly
flashed firmware.  This turns out to be insufficient on some machines.

This patch will wait up to five seconds, checking twice a second if the
device is available, and terminating early if it is.

Change-Id: Ib59d7f5ebaf234a5c2508253236f6bc3ee96c83f
Signed-off-by: Jes B. Klinke <jbk@chromium.org>