[csr_utils_pkg,dv] Avoid bug on reset in csr_spinwait
I broke this task for some pwrmgr tests with commit 5e0f0fe. The
problem is that I changed the fork/join_any so that it would complete
instantly on a reset and then use "disable fork" to kill the thing
that was waiting.

Unfortunately, the thing that got killed might have been in the middle
of a csr_rd and killing it doesn't allow the process to clean up
properly.

Go back to waiting more carefully but (in a bid to get *some* benefit
from the original rewrite) stop a wait between reads more quickly if
there is a reset.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>