[reggen] Represent pseudo-regs explicitly in MultiRegister
The existing code represented "the register to be replicated" with
self.reg and the new version explicitly represents the replicated
list (with self.pregs). Note that these aren't exactly "registers"
because won't necessarily all have distinct concrete registers

Note that we can also get rid of MultiRegister.count: now that we have
a list of pregs, this can actually be recovered from the length of
that list.

The behaviour should be unchanged, because this commit passes a list
of identical copies of the original register.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>