[kmac/dv] Add sideload_invalid test.
This commit adds the new sideload_invalid test. The goal of this test
is to increase coverage of the `StKeyMgrErrKeyNotValid` state of
the kmac_app FSM.

In the test, during a random FSM state, the sideload_key.valid signal
is forced to 0. The test checks whether the expected `ErrKeyNotValid`
error is seen.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>