[entropy_src/dv] Add missing 'value' argument to the csr_rd call
This change fixes an issue seen when trying to run the entropy_src
block using VCS. A call of csr_rd() function did not specify all
required arguments. This commit solves the issue by declaring a
variable (it's called unused for clarity that the value will not be
sampled elsewhere in code) and passing this 'unused' variable as the
value argument of the csr_rd call.

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>