[entropy_src, dv] Move assertion control code out of sequence
This change fixes a build issue seen when trying to run the entropy_src
using VCS simulator. The code to disable assertion from within sequences
using XMR doesnt compile.This is because VCS prohibits SV class inside
a package to have XMRs to code in module outside
The solution implemented in this commit moves the functions containing
assert_off tasks to the entropy_src_path_if.
These functions are then called as appropriate from the sequences using
a handle to the virtual interface

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>