[otbn, dv] Add missing ninja lib to python requirements
This fixes a run time issue seen when trying to run otbn tests using VCS
When running tests with vcs, the gen_binaries.py script looks for this
python library.
The fix implemented here adds ninja==1.11.1.1 requirement to
pyproject.toml file. The ./util/sh/scripts/gen-python-requirements.sh
script was run to update the python-requirements.txt file and the
updated python-requirements.txt is added to this commit along with the
source pyproject.toml file

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>