[verilator] Create verilator exec env aliases in //hw/top
To avoid a massive duplication, and as a first step toward a new
exec_env-less system, we will move all exec_env to //hw/top so that
they work for all tops. The easiest to change is verilator.
For now, only convert the test_rom_test to point to this new env.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>