[opentitanlib] Handle missing ownership seal in json
The ownership seal is a value applied to the ownership config by the
chip. It is meaningless to construct an owner config with a seal an
therefore, it should be permitted to omit it from the JSON
representation.

When decoding and printing an ownership config from a chip, however, it
is useful to print the value of the seal.

Signed-off-by: Chris Frantz <cfrantz@google.com>