[reggen] Explicitly register reset value in uvm_reg_base.sv.tpl
This will only matter for a multireg. Here, you might have many copies
of a register. These might be spread across fewer concrete registers,
but that doesn't matter too much for this discussion. Each concrete
register will be represented by the same UVM type. For example, the
pinmux registers have

    rand pinmux_reg_mio_pad_attr mio_pad_attr[47];

If one of those concrete registers should have a different reset value
then we're going to need to control that at a higher level than the
register class itself (pinmux_reg_mio_pad_attr, here).

The items in the array of register objects get created by
instantiate_register, which also takes a "reg" argument for the Python
register object that represents this particular item. With the
previous commits, we now have the relevant value as reg.resval, and
this value might be different for different items.

Explicitly tell UVM to set the reset value here. It will usually match
the one that got set in the uvm_register constructor, but this makes
it possible to override a single item.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>