[tlul,rtl] Remove a term in tlul_adapter_reg's instr_error signal
The instr_error signal is only used in an OR with tl_err. If the
instr_type field is invalid then u_err (a tlul_err instance) will
cause tl_err to be high (because it's equivalent check will fail). So
there's no need to test it explicitly here.

Note that the instr_error signal is still worth having, because it
allows us to spot when someone does a TL transaction whose user bits
claim to be of "instruction type" but where en_ifetch_i is false.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>