[sival,kmac] Add kmac_error_conditions_test
This adds the new `kmac_error_conditions_test` test that is
defined in the `chip_sw_kmac_error_conditions` chip level test.

`ErrSwIssuedCmdInAppActive` could not be implemented in SW as
triggering a SW command while the app interface is active is
not possible.

For the same reason, the `ErrSwPushedMsgFifo` was adapted.
Now, in this test point, a message is written into the message
FIFO before issuing the START command. This triggers the same
error code.

Closes #21413.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>