[dvsim] Add typing annotations to utils.py
This adds annotations to all functions except for
find_and_substitute_wildcards. That function is a bit complicated and
the recursion is much more complicated than it needs to be: we'll sort
out the implementation and types there separately.

No functional change with the other hunks, but this should allow mypy
to tell us something helpful for unrelated changes in dvsim.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>