[dvsim] Add lots of types to Testplan.py
This was rather an unsatisfying exercise, because the underlying
Python seems to fundamentally have rather dynamic typing. In
particular, the Testplan.progress object is represented as a nested
dict, with a type that changes through the run(!) This commit adds a
note about that (mostly as a reminder to tidy it up as a follow-up).

The other big change is adding various "casting" functions (as_string,
as_string_map, as_list, as_path_list, as_string_list,
as_string_map_list) that do the dynamic typing and fail with slightly
more helpful error messages if something gets unstuck.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>