[dvsim] Make Mode subclass a bit simpler
The "cunning design" that I'd originally made worked by passing a
type_name argument to base class constructor to tell it what is being
built. This was a bit more complicated than just having a class
variable, and confuses mypy when we want to pass the type as a
variable to represent the constructor.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>