[dv, dvsim] Use file descriptors to share logfile across processes
For non-interactive jobs, the Nclauncher script writes some initial
message into the log file and then forks off a subprocess to initiate
the call to the nc scheduler with the same log file pointer for stdout
logging. When multiple jobs are run in parallel and the log file resides
in a NFS, this seems to cause some clobbering of file handles causing
the subprocess to error with a cryptic error message saying socket
address already in use. Using a file descriptor can alleviate this
problem as they work at a lower level of abstraction and are managed by
the kernel and can handle open files across both the parent and child
processes.The file descriptor approach is now also extended to the
interactive mode which further allows removing redundant code by sharing
same logic between interactive and non-interactive modes

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>